+++ carp_mobile_sensing/lib/domain/device_info.dart
@@ -7,48 +7,50 @@
 
 part of domain;
 
-//TODO - make this a singleton class -- e.g. Device().deviceID
-
 /// Provides (static) information about the local device.
 ///
 /// This class is a singleton that one time access the information from the
 /// local device to be used in the sensing framework.
 class Device {
-  /// Device info about the device from which this datum were collected from.
-  static final DeviceInfoPlugin deviceInfoPlugin = DeviceInfoPlugin();
+  final DeviceInfoPlugin _deviceInfoPlugin = DeviceInfoPlugin();
 
-  static String platform;
-  static String hardware;
-  static String deviceID;
-  static String deviceName;
-  static String deviceManufacturer;
-  static String deviceModel;
-  static String operatingSystem;
-  static String sdk;
-  static String release;
+  static final Device _instance = Device._();
+  Device._() {
+    _getDeviceInfo();
+  }
+
+  factory Device() => _instance;
+
+  String platform;
+  String hardware;
+  String deviceID;
+  String deviceName;
+  String deviceManufacturer;
+  String deviceModel;
+  String operatingSystem;
+  String sdk;
+  String release;
 
   /// The device info for this device.
-  static Map<String, dynamic> deviceData = <String, dynamic>{};
+  Map<String, dynamic> deviceData = <String, dynamic>{};
 
   /// Get the device info using the [DeviceInfoPlugin].
-  static void getDeviceInfo() async {
+  void _getDeviceInfo() async {
     Map<String, dynamic> _deviceData;
 
     try {
       if (Platform.isAndroid) {
-        deviceData = _readAndroidBuildData(await deviceInfoPlugin.androidInfo);
+        deviceData = _readAndroidBuildData(await _deviceInfoPlugin.androidInfo);
       } else if (Platform.isIOS) {
-        deviceData = _readIosDeviceInfo(await deviceInfoPlugin.iosInfo);
+        deviceData = _readIosDeviceInfo(await _deviceInfoPlugin.iosInfo);
       }
     } on PlatformException {
-      deviceData = <String, dynamic>{
-        'Error:': 'Failed to get platform version.'
-      };
+      deviceData = <String, dynamic>{'Error:': 'Failed to get platform version.'};
     }
     deviceData = _deviceData;
   }
 
-  static Map<String, dynamic> _readAndroidBuildData(AndroidDeviceInfo info) {
+  Map<String, dynamic> _readAndroidBuildData(AndroidDeviceInfo info) {
     platform = 'Android';
     hardware = info.hardware;
     deviceID = info.id;
@@ -89,7 +91,7 @@ class Device {
     };
   }
 
-  static Map<String, dynamic> _readIosDeviceInfo(IosDeviceInfo info) {
+  Map<String, dynamic> _readIosDeviceInfo(IosDeviceInfo info) {
     platform = 'IOS';
     hardware = info.utsname.machine;
     deviceID = info.identifierForVendor;